requirements.txt

django-bootstrap-v5~=1.0
crispy-bootstrap5~=2023.10
coverage~=7.4
python-dotenv~=1.0
dj-database-url==2.1.0
django-crispy-forms==2.1
Django==4.2.7
gunicorn==21.2.0
Pillow==10.1.0
psycopg2==2.9.7
whitenoise==6.6.0